Time: 257 ms (21.15%), Space: 162.9 MB (22.17%) - LeetHub

+++ 3761-minimum-absolute-distance-between-mirror-pairs/3761-minimum-absolute-distance-between-mirror-pairs.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minMirrorPairDistance(vector<int>& nums) {
-        using ll = long long;
+        using ll = int;
         map<ll, ll> mp;
         auto rev = [&](ll x)
         {
